ftprintf mollinette

+++ course/ft_printf/project/ft_putnbr.c
@@ -1,38 +1,38 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
+/*   ft_putnbr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: bde-albu <bde-albu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/04 08:37:02 by bde-albu          #+#    #+#             */
-/*   Updated: 2025/04/04 08:51:29 by bde-albu         ###   ########.fr       */
+/*   Updated: 2025/04/10 10:14:02 by bde-albu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
 #include "ft_printf.h"
+#include <unistd.h>
 
-size_t	print(long n, int fd)
+int	print(long n, int fd)
 {
 	char	num;
-    size_t  length;
+	int		length;
 
-    length = 0;
+	length = 0;
 	if (n >= 10)
 		length += print(n / 10, fd);
 	num = (n % 10) + '0';
 	write(fd, &num, 1);
-    return (length + 1);
+	return (length + 1);
 }
 
-size_t	ft_putnbr(int n, int fd)
+int	ft_putnbr(int n, int fd)
 {
 	long	long_n;
-    size_t  len;
+	int		len;
 
 	long_n = n;
-    len = 0;
+	len = 0;
 	if (long_n == 0)
 	{
 		write(fd, "0", 1);
@@ -40,7 +40,6 @@ size_t	ft_putnbr(int n, int fd)
 	}
 	if (long_n < 0)
 		long_n *= -1;
-	
 	len += print(long_n, fd);
-    return (len);
+	return (len);
 }
